refactor(routes): promote helpers to first-class routes layout
- Move helpers/fastapi.py → fastapi_routes.py
- Move helpers/_proto_schema.py → _proto_schema.py
- Move helpers/jsonrpc.py → _jsonrpc_schema.py (renamed)
- Remove helpers/ subpackage
- Use Sequence[BaseRoute] instead of list[BaseRoute] for covariance
- Fix ty invalid-assignment error on app.openapi override
