+++ src/a2a/server/routes/__init__.py
@@ -5,7 +5,7 @@ from a2a.server.routes.common import (
     DefaultServerCallContextBuilder,
     ServerCallContextBuilder,
 )
-from a2a.server.routes.helpers import add_a2a_routes_to_fastapi
+from a2a.server.routes.fastapi_routes import add_a2a_routes_to_fastapi
 from a2a.server.routes.jsonrpc_routes import create_jsonrpc_routes
 from a2a.server.routes.rest_routes import create_rest_routes
