--- src/a2a/server/routes/helpers/__init__.py
@@ -1,6 +1,0 @@
-from a2a.server.routes.helpers.fastapi import add_a2a_routes_to_fastapi
-
-
-__all__ = [
-    'add_a2a_routes_to_fastapi',
-]
